Feat:  프론트와 백 연결 (#129)
- 프론트와 백 연결

+++ src/main/resources/static/js/viewJobSearch.js
@@ -2,34 +2,41 @@
 /* 본문 하단 버튼 */
 /***************/
 
-const editBtn = document.querySelector(".edit_btn");
-const deleteBtn = document.querySelector(".delete_btn");
-const reportBtn = document.querySelector(".report");
+if (
+  document.querySelector(".edit_btn") &&
+  document.querySelector(".delete_btn")
+) {
+  const editBtn = document.querySelector(".edit_btn");
+  const deleteBtn = document.querySelector(".delete_btn");
+  editBtn.addEventListener("click", editPost);
+  deleteBtn.addEventListener("click", openDeleteConfirm);
+  function editPost() {
+    window.location.href = "/s/api/jobSearch/updateForm/" + jobId;
+  }
 
-editBtn.addEventListener("click", editPost);
-deleteBtn.addEventListener("click", openDeleteConfirm);
-reportBtn.addEventListener("click", reportJob);
-
-function editPost() {
-  window.location.href = "writeJobSearch.html";
+  // 삭제확인창 열기
+  function openDeleteConfirm() {
+    let deleteConfirm = document.querySelector(".delete_confirm");
+    deleteConfirm.style.display = "flex";
+  }
+  // 삭제버튼들 연결
+  const confirmDelete = document.querySelector("#confirmDelete");
+  confirmDelete.addEventListener("click", () => {
+    deleteJob();
+  });
+  const closeDelete = document.querySelector("#closeDelete");
+  closeDelete.addEventListener("click", () => {
+    let deleteConfirm = document.querySelector(".delete_confirm");
+    deleteConfirm.style.display = "none";
+  });
 }
+const jobReportBtn = document.querySelector(".jobReportBtn");
+jobReportBtn.addEventListener("click", reportJob);
 
-// 삭제확인창 열기
-function openDeleteConfirm() {
-  let deleteConfirm = document.querySelector(".delete_confirm");
-  deleteConfirm.style.display = "flex";
-}
+// const commentReportBtn = document.querySelector(".commentReport");
+// commentReportBtn.addEventListener("click", reportComment);
 
-// 삭제버튼들 연결
-const confirmDelete = document.querySelector("#confirmDelete");
-confirmDelete.addEventListener("click", () => {
-  deleteJob();
-});
-const closeDelete = document.querySelector("#closeDelete");
-closeDelete.addEventListener("click", () => {
-  let deleteConfirm = document.querySelector(".delete_confirm");
-  deleteConfirm.style.display = "none";
-});
+let jobId = $("#jobId").val();
 
 /*******/
 /* 댓글 */
@@ -52,7 +59,7 @@ function submitFeedback(e) {
   // get feedback
   const commentForm = comment.value;
   const currentTime = new Date();
-  // if inputs are not empty
+  // if inputs are not empt삭
   if (userForm && commentForm !== "") {
     // create new feedback
     newFeedback = {
@@ -96,21 +103,26 @@ function addFeedback(item) {
                     <p class="nickname">
                         ${item.userName}
                     </p>
-                    <p class="created_date">${item.createdDate}</p>
+                    <p class="created_date"><fmt:formatDate value="${item.createdDate}" pattern="yy.MM.dd kk:mm" type="date" /></p>
                 </div>
             </div>
             <div class="comment_btns">
                 <button class='editBtn'>수정</button>
                 <button class='removeBtn'>삭제</button>
-                <button class='reportBtn'>신고하기</button>
+                <button class='commentReportBtn'>신고하기</button>
             </div>
         </div>
-        <textarea class="comment textarea" readonly>${item.userComment}</textarea>
-        
+        <textarea class="comment textarea" id="commentContent" readonly>${item.userComment}</textarea>
+
     `;
   // insert feedback into the list
   commentsCont.insertAdjacentElement("beforeend", div);
   activeCommentBtn();
+}
+
+// 댓글 창을 초기화 함수
+function resetForm() {
+  comment.value = "";
 }
 
 // 댓글 버튼 활성화
@@ -118,7 +130,7 @@ function activeCommentBtn() {
   // 각 버튼 요소 선택
   const editButtons = document.querySelectorAll(".editBtn");
   const removeButtons = document.querySelectorAll(".removeBtn");
-  const reportButtons = document.querySelectorAll(".reportBtn");
+  const reportButtons = document.querySelectorAll(".commentReportBtn");
 
   // 각 버튼과 함수 연결
   editButtons.forEach((editButton) => {
@@ -153,6 +165,25 @@ function saveComment() {
   textarea.readOnly = true;
   textarea.style.border = "none";
 
+  let data = {
+    jobId: $("#jobId").val(),
+    commentId: commentCard.querySelector(".commentId").value,
+    commentContent: textarea.value,
+    userId: $("#userId").val()
+  };
+
+  $.ajax("/s/api/comment/update", {
+    type: "PUT",
+    data: JSON.stringify(data),
+    dataType: "json",
+    headers: {
+      "Content-Type": "application/json; charset=utf-8"
+    }
+  }).done((res) => {
+    if (res.code == 1) {
+      location.reload();
+    }
+  });
   const editBtn = commentCard.querySelector(".editBtn");
   editBtn.innerText = "수정";
   editBtn.removeEventListener("click", saveComment);
@@ -162,27 +193,46 @@ function saveComment() {
 // 댓글 삭제
 function removeComment() {
   const commentCard = this.closest(".comment_card"); // 삭제 버튼의 부모 comment_card 요소 찾기
-  commentsCont.removeChild(commentCard);
-}
+  if (confirm("정말 삭제하시겠습니까?")) {
+    const commentCard = this.closest(".comment_card"); // 저장 버튼의 부모 comment_card 요소 찾기
 
-// 구인글 신고하기
-$(".jobReportBtn").click(() => {
-  reportJob();
-});
+    let data = {
+      commentId: commentCard.querySelector(".commentId").value,
+      jobId: $("#jobId").val(),
+      userId: $("#userId").val()
+    };
+
+    $.ajax("/s/api/comment/delete", {
+      type: "DELETE",
+      dataType: "json",
+      data: JSON.stringify(data),
+      headers: {
+        "Content-Type": "application/json"
+      }
+    }).done((res) => {
+      if (res.code == 1) {
+        location.reload();
+      } else {
+        alert(res.msg);
+        return false;
+      }
+    });
+  }
+}
 
 //  구인글 신고하기
 function reportJob() {
   let jobId = $("#jobId").val();
   let userId = $("#jobUserId").val();
 
-  var nWidth = "500";
-  var nHeight = "600";
+  var nWidth = "700";
+  var nHeight = "900";
   var xPos = document.body.clientWidth / 2 - nWidth / 2;
   xPos += window.screenLeft; //듀얼 모니터
   var yPos = screen.availHeight / 2 - nHeight / 2;
 
   window.open(
-    "/s/api/reportForm/" + jobId + "/" + userId,
+    "/s/api/reportFormJob/" + jobId + "/" + userId,
     "신고하기",
     "width=" +
       nWidth +
@@ -197,15 +247,17 @@ function reportJob() {
 }
 
 //  댓글 신고하기
-function reportComment() {
-  var nWidth = "500";
-  var nHeight = "600";
+function reportComment(commentId) {
+  let userId = $("#userId").val();
+
+  var nWidth = "700";
+  var nHeight = "900";
   var xPos = document.body.clientWidth / 2 - nWidth / 2;
   xPos += window.screenLeft; //듀얼 모니터
   var yPos = screen.availHeight / 2 - nHeight / 2;
 
   window.open(
-    "/s/api/reportForm",
+    "/s/api/reportFormComment/" + commentId + "/" + userId,
     "신고하기",
     "width=" +
       nWidth +
@@ -218,6 +270,7 @@ function reportComment() {
       ",toolbar=no"
   );
 }
+
 /************************/
 /* textarea 높이 조절 함수 */
 /************************/
@@ -318,8 +371,40 @@ function deleteJob() {
     }
   }).done((res) => {
     if (res.code == 1) {
-      alert(res.msg);
       location.href = "/jobSearch/mainForm";
+    } else {
+      alert(res.msg);
+      return false;
+    }
+  });
+}
+
+// 댓글 작성
+$("#commentWriteBtn").click(() => {
+  write();
+});
+
+function write() {
+  let data = {
+    commentContent: $("#commentContent").val(),
+    jobId: $("#jobId").val(),
+    userId: $("#userId").val()
+  };
+  if (data.commentContent.length < 1) {
+    alert("댓글을 입력해주세요.");
+    return;
+  }
+
+  $.ajax("/s/api/comment/write", {
+    type: "POST",
+    dataType: "json",
+    data: JSON.stringify(data),
+    headers: {
+      "Content-Type": "application/json"
+    }
+  }).done((res) => {
+    if (res.code == 1) {
+      location.href = "/s/api/jobSearch/detailForm/" + data.jobId;
     } else {
       alert(res.msg);
       return false;
